homepage layout rewrite part 5
