+++ src/components/Footer.jsx
@@ -13,33 +13,48 @@ const Footer = () => {
     <footer className='footer'>
       <Container className='mobile-footer-container'>
         {onDesktop ? (
-          // 桌面端布局
           <Row className='d-flex desktop-layout'>
-            <Col md={3} className='footer-column logo-contact'>
+            <Row className='footer-column logo-contact'>
               <div className='d-flex align-items-center'>
                 <img
                   src='/footer_logo.png'
                   alt='Logo'
                   className='footer-logo me-3'
                 />
+              </div>
+            </Row>
+              <div>
                 <Link to={`/merchant/360-media-promotion-service`}>
                   <Button className='update-function-btn'>加入我们</Button>
                 </Link>
+                <Link to={`/about-us`}>
+                  <Button className='update-function-btn'>关于我们</Button>
+                </Link>
               </div>
-              <div className='contact-info mt-3'>
+
+              <div>
+                <Link to={`/merchant/360-media-promotion-service`}>
+                  <Button className='update-function-btn'>活动</Button>
+                </Link>
+                <Link to={`/merchant/360-media-promotion-service`}>
+                  <Button className='update-function-btn'>新闻</Button>
+                </Link>
+              </div>
+            <Row>
                 <Row>
                   <Col
-                    xs={1}
-                    className='d-flex justify-content-center align-items-center'
-                  >
-                    <i className='bi bi-pin-map-fill'></i>
-                  </Col>
-                  <Col>
-                    <p>
-                      L2 171 La Trobe Street <br /> Melbourne VIC 3000
-                    </p>
+                      xs={1}
+                      className='d-flex justify-content-center align-items-center'
+                    >
+                      <i className='bi bi-pin-map-fill'></i>
+                    </Col>
+                    <Col>
+                      <p>
+                        L2 171 La Trobe Street <br /> Melbourne VIC 3000
+                      </p>
                   </Col>
                 </Row>
+
                 <Row>
                   <Col
                     xs={1}
@@ -55,6 +70,7 @@ const Footer = () => {
                     </p>
                   </Col>
                 </Row>
+
                 <Row>
                   <Col
                     xs={1}
@@ -66,18 +82,9 @@ const Footer = () => {
                     <p>info@do360.com</p>
                   </Col>
                 </Row>
-              </div>
-            </Col>
-
-            <Col md={6}></Col>
-
-            <Col md={3} className='footer-column qr-social text-end'>
-              <img src='/QR_JohnDu.png' alt='QR Code' className='qr-code' />
-              <p>{t("footer_scan")}</p>
-            </Col>
+            </Row>
           </Row>
         ) : (
-          // 移动端：将 4 个元素分成 2 排
           <div>
             <Image
               src='/footer_logo.png'
@@ -85,8 +92,23 @@ const Footer = () => {
               className='mobile-footer-logo'
               fluid
             />
+            <div>
+              <Link to={`/merchant/360-media-promotion-service`}>
+                <Button className='update-function-btn'>加入我们</Button>
+              </Link>
+              <Link to={`/merchant/360-media-promotion-service`}>
+                <Button className='update-function-btn'>关于我们</Button>
+              </Link>
+            </div>
+            <div>
+              <Link to={`/merchant/360-media-promotion-service`}>
+                <Button className='update-function-btn'>活动</Button>
+              </Link>
+              <Link to={`/merchant/360-media-promotion-service`}>
+                <Button className='update-function-btn'>新闻</Button>
+              </Link>
+            </div>
             <Row className='footer-contact-info'>
-              {/* 第一排：地址 & 电话 */}
               <Col xs={6} className='d-flex align-items-center'>
                 <i className='bi bi-pin-map-fill'></i>
                 <p className='ms-2'>
@@ -102,15 +124,9 @@ const Footer = () => {
                 </p>
               </Col>
 
-              {/* 第二排：邮箱 & 加入我们 */}
               <Col xs={6} className='d-flex align-items-center mt-2'>
                 <i className='bi bi-mailbox2'></i>
                 <p className='ms-2'>info@do360.com</p>
-              </Col>
-              <Col xs={6} className='d-flex align-items-center mt-2'>
-                <Link to={`/merchant/360-media-promotion-service`}>
-                  <Button className='update-function-btn'>加入我们</Button>
-                </Link>
               </Col>
             </Row>
           </div>
